refactor(*): refactor everything
