+++ README.md
@@ -1,15 +1,3 @@
-# TypeScript Template
-Template for TypeScript projects
+# QuickMongo
 
-# Included
-- eslint
-- unit test
-- prettier
-- typescript declaration files bundler
-- es modules generator
-- husky pre-commit
-- github actions:
-  - CodeQL
-  - ESLint
-  - Auto `npm publish` on release creation
-  - Stale action
+Quick Mongodb wrapper for beginners.
