+++ package.json
@@ -1,7 +1,7 @@
 {
-  "name": "@devsnowflake/typescript-template",
-  "version": "1.0.0",
-  "description": "Template for TypeScript projects",
+  "name": "quickmongo",
+  "version": "4.0.0-dev",
+  "description": "Quick mongodb wrapper for beginners.",
   "main": "dist/index.js",
   "module": "dist/index.mjs",
   "types": "dist/index.d.ts",
@@ -18,26 +18,31 @@
     "build:esm": "gen-esm-wrapper ./dist/index.js ./dist/index.mjs",
     "lint": "eslint src --ext .ts",
     "lint:fix": "eslint src --ext .ts --fix",
-    "bumpdeps": "npm-check-updates -u",
     "format": "prettier --write src/**/*.{ts,js,json,yaml,yml}",
     "prepublishOnly": "rollup-type-bundler -e stream",
     "test": "jest"
   },
   "repository": {
     "type": "git",
-    "url": "git+https://github.com/DevSnowflake/typescript-template.git"
+    "url": "git+https://github.com/DevSnowflake/quickmongo.git"
   },
   "keywords": [
-    "typescript-template",
-    "template",
-    "project"
+    "quickmongo",
+    "mongodb",
+    "mongoose",
+    "schema",
+    "api",
+    "database",
+    "quick.db",
+    "enmap",
+    "endb"
   ],
-  "author": "Snowflake Studio ❄️",
+  "author": "DevAndromeda",
   "license": "MIT",
   "bugs": {
-    "url": "https://github.com/DevSnowflake/typescript-template/issues"
+    "url": "https://github.com/DevSnowflake/quickmongo/issues"
   },
-  "homepage": "https://github.com/DevSnowflake/typescript-template#readme",
+  "homepage": "https://github.com/DevSnowflake/quickmongo#readme",
   "devDependencies": {
     "@babel/core": "^7.15.0",
     "@babel/preset-env": "^7.15.0",
@@ -52,7 +57,6 @@
     "husky": "^7.0.1",
     "jest": "^27.0.6",
     "mongodb": "^4.1.1",
-    "npm-check-updates": "^11.8.3",
     "prettier": "^2.3.2",
     "rimraf": "^3.0.2",
     "ts-node": "^10.1.0",
