+++ src/index.ts
@@ -1,0 +1,2 @@
+// eslint-disable-next-line @typescript-eslint/no-var-requires
+export const version = require("../package.json").version;
